Retrieve table schemas from JSON
Remove hardcoded tables
Presort tables
Sort Table.Indexes by name
