--- src/db/table_definitions.go
@@ -1,1850 +1,0 @@
-package db
-
-import (
-	"path/filepath"
-
-	"github.com/Eve-of-Darkness/db-public/src/utils"
-	"golang.org/x/exp/slices"
-
-	"encoding/json"
-	"os"
-)
-
-func GetAllTables() []Table {
-	allTables := make([]Table, 0)
-	t := NewTable("Ability")
-	t.AddPrimary("AbilityID", "int(11)").NotNullable()
-	t.AddUnique("KeyName", "varchar(100)").NotNullable()
-	t.Add("Name", "varchar(255)").NotNullable()
-	t.Add("InternalID", "int(11)").NotNullable()
-	t.Add("Description", "text").NotNullable()
-	t.Add("IconID", "int(11)").NotNullable()
-	t.Add("Implementation", "varchar(255)")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 241
-	allTables = append(allTables, *t)
-
-	t = NewTable("Account")
-	t.AddPrimary("Name", "varchar(255)").NotNullable()
-	t.Add("Password", "text").NotNullable()
-	t.Add("CreationDate", "datetime").NotNullable()
-	t.Add("LastLogin", "datetime")
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("PrivLevel", "int(10) unsigned").NotNullable()
-	t.Add("Status", "int(11)").NotNullable()
-	t.Add("Mail", "text")
-	t.AddWithIndex("LastLoginIP", "varchar(255)")
-	t.Add("LastClientVersion", "text")
-	t.Add("Language", "text")
-	t.Add("IsMuted", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Account_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("AccountXCustomParam")
-	t.AddWithIndex("Name", "varchar(255)").NotNullable()
-	t.AddWithIndex("KeyName", "varchar(100)").NotNullable()
-	t.Add("Value", "varchar(255)")
-	t.AddPrimary("CustomParamID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("Appeal")
-	t.AddWithIndex("Name", "varchar(255)").NotNullable()
-	t.AddWithIndex("Account", "varchar(255)").NotNullable()
-	t.Add("Severity", "int(11)").NotNullable()
-	t.Add("Status", "text").NotNullable()
-	t.Add("Timestamp", "text").NotNullable()
-	t.Add("Text", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Appeal_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Area")
-	t.Add("TranslationId", "text")
-	t.Add("Description", "text").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Radius", "int(11)").NotNullable()
-	t.Add("Region", "smallint(5) unsigned").NotNullable()
-	t.Add("ClassType", "text")
-	t.Add("CanBroadcast", "tinyint(1)").NotNullable()
-	t.Add("Sound", "tinyint(3) unsigned").NotNullable()
-	t.Add("CheckLOS", "tinyint(1)").NotNullable()
-	t.Add("Points", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Area_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("ArtifactBonus")
-	t.Add("ArtifactID", "text").NotNullable()
-	t.Add("BonusID", "int(11)").NotNullable()
-	t.Add("Level", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("ArtifactBonus_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Artifact")
-	t.Add("ArtifactID", "text").NotNullable()
-	t.Add("EncounterID", "text").NotNullable()
-	t.Add("QuestID", "text").NotNullable()
-	t.Add("Zone", "text").NotNullable()
-	t.Add("ScholarID", "text").NotNullable()
-	t.Add("ReuseTimer", "int(11)").NotNullable()
-	t.Add("XPRate", "int(11)").NotNullable()
-	t.Add("BookID", "text").NotNullable()
-	t.Add("BookModel", "int(11)").NotNullable()
-	t.Add("Scroll1", "text").NotNullable()
-	t.Add("Scroll2", "text").NotNullable()
-	t.Add("Scroll3", "text").NotNullable()
-	t.Add("Scroll12", "text").NotNullable()
-	t.Add("Scroll13", "text").NotNullable()
-	t.Add("Scroll23", "text").NotNullable()
-	t.Add("ScrollModel1", "int(11)").NotNullable()
-	t.Add("ScrollModel2", "int(11)").NotNullable()
-	t.Add("ScrollLevel", "int(11)").NotNullable()
-	t.Add("MessageUse", "text").NotNullable()
-	t.Add("MessageCombineScrolls", "text").NotNullable()
-	t.Add("MessageCombineBook", "text").NotNullable()
-	t.Add("MessageReceiveScrolls", "text").NotNullable()
-	t.Add("MessageReceiveBook", "text").NotNullable()
-	t.Add("Credit", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Artifact_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("ArtifactXItem")
-	t.Add("ArtifactID", "text").NotNullable()
-	t.Add("ItemID", "text").NotNullable()
-	t.Add("Version", "text").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("ArtifactXItem_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("AuditEntry")
-	t.Add("AuditTime", "datetime").NotNullable()
-	t.Add("AccountID", "text")
-	t.Add("RemoteHost", "text")
-	t.Add("AuditType", "int(11)").NotNullable()
-	t.Add("AuditSubtype", "int(11)").NotNullable()
-	t.Add("OldValue", "text")
-	t.Add("NewValue", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("AuditEntry_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("AutoXMLUpdate")
-	t.AddPrimary("AutoXMLUpdateID", "int(11)").NotNullable()
-	t.AddWithIndex("FilePackage", "varchar(255)").NotNullable()
-	t.AddWithIndex("FileHash", "varchar(255)").NotNullable()
-	t.Add("LoadResult", "varchar(255)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 4
-	allTables = append(allTables, *t)
-
-	t = NewTable("Ban")
-	t.Add("Author", "text").NotNullable()
-	t.AddWithIndex("Type", "varchar(255)").NotNullable()
-	t.AddWithIndex("Ip", "varchar(255)").NotNullable()
-	t.AddWithIndex("Account", "varchar(255)").NotNullable()
-	t.Add("DateBan", "datetime").NotNullable()
-	t.Add("Reason", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Ban_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Battleground")
-	t.Add("RegionID", "smallint(5) unsigned").NotNullable()
-	t.Add("MinLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("MaxLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("MaxRealmLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Battleground_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("BindPoint")
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Radius", "smallint(5) unsigned").NotNullable()
-	t.Add("Region", "int(11)").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("BindPoint_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("BugReport")
-	t.AddPrimary("ID", "int(11)").NotNullable()
-	t.Add("Message", "text").NotNullable()
-	t.Add("Submitter", "text").NotNullable()
-	t.Add("DateSubmitted", "datetime").NotNullable()
-	t.Add("ClosedBy", "text").NotNullable()
-	t.Add("DateClosed", "datetime").NotNullable()
-	t.Add("Category", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("BugReport_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("CharacterXDataQuest")
-	t.AddPrimary("ID", "int(11)").NotNullable()
-	t.Add("Character_ID", "varchar(100)").NotNullable()
-	t.Add("DataQuestID", "int(11)").NotNullable()
-	t.Add("Step", "smallint(6)").NotNullable()
-	t.Add("Count", "smallint(6)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	index := new(Index)
-	index.Name = "I_CharacterXDataQuest_Character_ID"
-	index.Column = "DataQuestID"
-	t.Indexes = append(t.Indexes, index)
-	index2 := new(Index)
-	index2.Name = "I_CharacterXDataQuest_Character_ID"
-	index2.Column = "Character_ID"
-	t.Indexes = append(t.Indexes, index2)
-	allTables = append(allTables, *t)
-
-	t = NewTable("CharacterXMasterLevel")
-	t.AddWithIndex("Character_ID", "varchar(255)").NotNullable()
-	t.Add("MLLevel", "int(11)").NotNullable()
-	t.Add("MLStep", "int(11)").NotNullable()
-	t.Add("StepCompleted", "tinyint(1)").NotNullable()
-	t.Add("ValidationDate", "datetime")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("CharacterXMasterLevel_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("CharacterXOneTimeDrop")
-	t.AddWithIndex("CharacterID", "varchar(100)").NotNullable()
-	t.AddWithIndex("ItemTemplateID", "varchar(100)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("CharacterXOneTimeDrop_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("ClassXRealmAbility")
-	t.Add("CharClass", "int(11)").NotNullable()
-	t.Add("AbilityKey", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("ClassXRealmAbility_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("ClassXSpecialization")
-	t.AddPrimary("ClassXSpecializationID", "int(11)").NotNullable()
-	t.AddWithIndex("ClassID", "int(11)").NotNullable()
-	t.AddWithIndex("SpecKeyName", "varchar(100)").NotNullable()
-	t.AddWithIndex("LevelAcquired", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 538
-	allTables = append(allTables, *t)
-
-	t = NewTable("CraftedItem")
-	t.AddPrimary("CraftedItemID", "varchar(255)").NotNullable()
-	t.AddWithIndex("Id_nb", "varchar(255)").NotNullable()
-	t.Add("CraftingLevel", "int(11)").NotNullable()
-	t.Add("CraftingSkillType", "int(11)").NotNullable()
-	t.Add("MakeTemplated", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("CraftedItem_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("CraftedXItem")
-	t.AddWithIndex("CraftedItemId_nb", "varchar(255)").NotNullable()
-	t.Add("IngredientId_nb", "text").NotNullable()
-	t.Add("Count", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("CraftedXItem_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DataQuest")
-	t.AddPrimary("ID", "int(11)").NotNullable()
-	t.Add("Name", "varchar(255)").NotNullable()
-	t.Add("StartType", "tinyint(3) unsigned").NotNullable()
-	t.Add("StartName", "varchar(100)").NotNullable()
-	t.Add("StartRegionID", "smallint(5) unsigned").NotNullable()
-	t.Add("AcceptText", "varchar(100)")
-	t.Add("Description", "text")
-	t.Add("SourceName", "text")
-	t.Add("SourceText", "text")
-	t.Add("StepType", "text")
-	t.Add("StepText", "text")
-	t.Add("StepItemTemplates", "text")
-	t.Add("AdvanceText", "text")
-	t.Add("TargetName", "text")
-	t.Add("TargetText", "text")
-	t.Add("CollectItemTemplate", "text")
-	t.Add("MaxCount", "smallint(5) unsigned").NotNullable()
-	t.Add("MinLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("MaxLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("RewardMoney", "text")
-	t.Add("RewardXP", "text")
-	t.Add("RewardCLXP", "text")
-	t.Add("RewardRP", "text")
-	t.Add("RewardBP", "text")
-	t.Add("OptionalRewardItemTemplates", "text")
-	t.Add("FinalRewardItemTemplates", "text")
-	t.Add("FinishText", "text")
-	t.Add("QuestDependency", "text")
-	t.Add("AllowedClasses", "varchar(200)")
-	t.Add("ClassType", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("DBHouseCharsXPerms")
-	t.Add("HouseNumber", "int(11)").NotNullable()
-	t.Add("PermissionType", "int(11)").NotNullable()
-	t.Add("TargetName", "text").NotNullable()
-	t.Add("DisplayName", "text").NotNullable()
-	t.Add("PermissionLevel", "int(11)").NotNullable()
-	t.Add("CreationTime", "datetime").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("DBHouseCharsXPerms_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DBHousePermissions")
-	t.Add("PermissionLevel", "int(11)").NotNullable()
-	t.Add("HouseNumber", "int(11)").NotNullable()
-	t.Add("CanEnterHouse", "tinyint(1)").NotNullable()
-	t.Add("Vault1", "tinyint(3) unsigned").NotNullable()
-	t.Add("Vault2", "tinyint(3) unsigned").NotNullable()
-	t.Add("Vault3", "tinyint(3) unsigned").NotNullable()
-	t.Add("Vault4", "tinyint(3) unsigned").NotNullable()
-	t.Add("CanChangeExternalAppearance", "tinyint(1)").NotNullable()
-	t.Add("ChangeInterior", "tinyint(3) unsigned").NotNullable()
-	t.Add("ChangeGarden", "tinyint(3) unsigned").NotNullable()
-	t.Add("CanBanish", "tinyint(1)").NotNullable()
-	t.Add("CanUseMerchants", "tinyint(1)").NotNullable()
-	t.Add("CanUseTools", "tinyint(1)").NotNullable()
-	t.Add("CanBindInHouse", "tinyint(1)").NotNullable()
-	t.Add("ConsignmentMerchant", "tinyint(3) unsigned").NotNullable()
-	t.Add("CanPayRent", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("DBHousePermissions_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DBHouse")
-	t.AddPrimary("HouseNumber", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.AddWithIndex("RegionID", "smallint(5) unsigned").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.Add("CreationTime", "datetime")
-	t.Add("Name", "text")
-	t.Add("Model", "int(11)").NotNullable()
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("PorchRoofColor", "int(11)").NotNullable()
-	t.Add("PorchMaterial", "int(11)").NotNullable()
-	t.Add("RoofMaterial", "int(11)").NotNullable()
-	t.Add("DoorMaterial", "int(11)").NotNullable()
-	t.Add("WallMaterial", "int(11)").NotNullable()
-	t.Add("TrussMaterial", "int(11)").NotNullable()
-	t.Add("WindowMaterial", "int(11)").NotNullable()
-	t.Add("Rug1Color", "int(11)").NotNullable()
-	t.Add("Rug2Color", "int(11)").NotNullable()
-	t.Add("Rug3Color", "int(11)").NotNullable()
-	t.Add("Rug4Color", "int(11)").NotNullable()
-	t.Add("IndoorGuildBanner", "tinyint(1)").NotNullable()
-	t.Add("IndoorGuildShield", "tinyint(1)").NotNullable()
-	t.Add("OutdoorGuildBanner", "tinyint(1)").NotNullable()
-	t.Add("OutdoorGuildShield", "tinyint(1)").NotNullable()
-	t.Add("Porch", "tinyint(1)").NotNullable()
-	t.Add("OwnerID", "text")
-	t.Add("LastPaid", "datetime")
-	t.Add("KeptMoney", "bigint(20)").NotNullable()
-	t.Add("NoPurge", "tinyint(1)").NotNullable()
-	t.Add("GuildHouse", "tinyint(1)").NotNullable()
-	t.Add("GuildName", "text")
-	t.Add("HasConsignment", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("DBHouse_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("DBIndoorItem")
-	t.AddWithIndex("HouseNumber", "int(11)").NotNullable()
-	t.Add("Model", "int(11)").NotNullable()
-	t.Add("Position", "int(11)").NotNullable()
-	t.Add("Placemode", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("BaseItemID", "text").NotNullable()
-	t.Add("Color", "int(11)").NotNullable()
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("Rotation", "int(11)").NotNullable()
-	t.Add("Size", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("DBIndoorItem_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DBOutdoorItem")
-	t.AddWithIndex("HouseNumber", "int(11)").NotNullable()
-	t.Add("Model", "int(11)").NotNullable()
-	t.Add("Position", "int(11)").NotNullable()
-	t.Add("Rotation", "int(11)").NotNullable()
-	t.Add("BaseItemID", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("DBOutdoorItem_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DOLCharactersBackup")
-	t.Add("DOLCharacters_ID", "varchar(255)").NotNullable()
-	t.AddWithIndex("Name", "varchar(255)").NotNullable()
-	t.Add("DeleteDate", "datetime").NotNullable()
-	t.Add("HasGravestone", "tinyint(1)").NotNullable()
-	t.Add("GravestoneRegion", "int(11)").NotNullable()
-	t.Add("Constitution", "int(11)").NotNullable()
-	t.Add("Dexterity", "int(11)").NotNullable()
-	t.Add("Strength", "int(11)").NotNullable()
-	t.Add("Quickness", "int(11)").NotNullable()
-	t.Add("Intelligence", "int(11)").NotNullable()
-	t.Add("Piety", "int(11)").NotNullable()
-	t.Add("Empathy", "int(11)").NotNullable()
-	t.Add("Charisma", "int(11)").NotNullable()
-	t.Add("BountyPoints", "bigint(20)").NotNullable()
-	t.Add("RealmPoints", "bigint(20)").NotNullable()
-	t.Add("RealmLevel", "int(11)").NotNullable()
-	t.Add("Experience", "bigint(20)").NotNullable()
-	t.Add("MaxEndurance", "int(11)").NotNullable()
-	t.Add("Health", "int(11)").NotNullable()
-	t.Add("Mana", "int(11)").NotNullable()
-	t.Add("Endurance", "int(11)").NotNullable()
-	t.Add("Concentration", "int(11)").NotNullable()
-	t.AddWithIndex("AccountName", "varchar(255)").NotNullable()
-	t.Add("AccountSlot", "int(11)").NotNullable()
-	t.Add("CreationDate", "datetime").NotNullable()
-	t.Add("LastPlayed", "datetime")
-	t.Add("LastName", "text")
-	t.AddWithIndex("GuildID", "varchar(255)")
-	t.Add("Gender", "int(11)").NotNullable()
-	t.Add("Race", "int(11)").NotNullable()
-	t.Add("Level", "int(11)").NotNullable()
-	t.Add("Class", "int(11)").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("CreationModel", "int(11)").NotNullable()
-	t.Add("CurrentModel", "int(11)").NotNullable()
-	t.Add("Region", "int(11)").NotNullable()
-	t.Add("ActiveWeaponSlot", "tinyint(3) unsigned").NotNullable()
-	t.Add("Xpos", "int(11)").NotNullable()
-	t.Add("Ypos", "int(11)").NotNullable()
-	t.Add("Zpos", "int(11)").NotNullable()
-	t.Add("BindXpos", "int(11)").NotNullable()
-	t.Add("BindYpos", "int(11)").NotNullable()
-	t.Add("BindZpos", "int(11)").NotNullable()
-	t.Add("BindRegion", "int(11)").NotNullable()
-	t.Add("BindHeading", "int(11)").NotNullable()
-	t.Add("BindHouseXpos", "int(11)").NotNullable()
-	t.Add("BindHouseYpos", "int(11)").NotNullable()
-	t.Add("BindHouseZpos", "int(11)").NotNullable()
-	t.Add("BindHouseRegion", "int(11)").NotNullable()
-	t.Add("BindHouseHeading", "int(11)").NotNullable()
-	t.Add("DeathCount", "tinyint(3) unsigned").NotNullable()
-	t.Add("ConLostAtDeath", "int(11)").NotNullable()
-	t.Add("Direction", "int(11)").NotNullable()
-	t.Add("MaxSpeed", "int(11)").NotNullable()
-	t.Add("Copper", "int(11)").NotNullable()
-	t.Add("Silver", "int(11)").NotNullable()
-	t.Add("Gold", "int(11)").NotNullable()
-	t.Add("Platinum", "int(11)").NotNullable()
-	t.Add("Mithril", "int(11)").NotNullable()
-	t.Add("SerializedCraftingSkills", "text")
-	t.Add("SerializedAbilities", "text")
-	t.Add("SerializedSpecs", "text")
-	t.Add("SerializedRealmAbilities", "text")
-	t.Add("DisabledSpells", "text")
-	t.Add("DisabledAbilities", "text")
-	t.Add("SerializedFriendsList", "text")
-	t.Add("SerializedIgnoreList", "text")
-	t.Add("IsCloakHoodUp", "tinyint(1)").NotNullable()
-	t.Add("IsCloakInvisible", "tinyint(1)").NotNullable()
-	t.Add("IsHelmInvisible", "tinyint(1)").NotNullable()
-	t.Add("SpellQueue", "tinyint(1)").NotNullable()
-	t.Add("IsLevelSecondStage", "tinyint(1)").NotNullable()
-	t.Add("FlagClassName", "tinyint(1)").NotNullable()
-	t.Add("Advisor", "tinyint(1)").NotNullable()
-	t.Add("GuildRank", "smallint(5) unsigned").NotNullable()
-	t.Add("PlayedTime", "bigint(20)").NotNullable()
-	t.Add("DeathTime", "bigint(20)").NotNullable()
-	t.Add("RespecAmountAllSkill", "int(11)").NotNullable()
-	t.Add("RespecAmountSingleSkill", "int(11)").NotNullable()
-	t.Add("RespecAmountRealmSkill", "int(11)").NotNullable()
-	t.Add("RespecAmountDOL", "int(11)").NotNullable()
-	t.Add("RespecAmountChampionSkill", "int(11)").NotNullable()
-	t.Add("IsLevelRespecUsed", "tinyint(1)").NotNullable()
-	t.Add("RespecBought", "int(11)").NotNullable()
-	t.Add("SafetyFlag", "tinyint(1)").NotNullable()
-	t.Add("CraftingPrimarySkill", "int(11)").NotNullable()
-	t.Add("CancelStyle", "tinyint(1)").NotNullable()
-	t.Add("IsAnonymous", "tinyint(1)").NotNullable()
-	t.Add("CustomisationStep", "tinyint(3) unsigned").NotNullable()
-	t.Add("EyeSize", "tinyint(3) unsigned").NotNullable()
-	t.Add("LipSize", "tinyint(3) unsigned").NotNullable()
-	t.Add("EyeColor", "tinyint(3) unsigned").NotNullable()
-	t.Add("HairColor", "tinyint(3) unsigned").NotNullable()
-	t.Add("FaceType", "tinyint(3) unsigned").NotNullable()
-	t.Add("HairStyle", "tinyint(3) unsigned").NotNullable()
-	t.Add("MoodType", "tinyint(3) unsigned").NotNullable()
-	t.Add("UsedLevelCommand", "tinyint(1)").NotNullable()
-	t.Add("CurrentTitleType", "text")
-	t.Add("KillsAlbionPlayers", "int(11)").NotNullable()
-	t.Add("KillsMidgardPlayers", "int(11)").NotNullable()
-	t.Add("KillsHiberniaPlayers", "int(11)").NotNullable()
-	t.Add("KillsAlbionDeathBlows", "int(11)").NotNullable()
-	t.Add("KillsMidgardDeathBlows", "int(11)").NotNullable()
-	t.Add("KillsHiberniaDeathBlows", "int(11)").NotNullable()
-	t.Add("KillsAlbionSolo", "int(11)").NotNullable()
-	t.Add("KillsMidgardSolo", "int(11)").NotNullable()
-	t.Add("KillsHiberniaSolo", "int(11)").NotNullable()
-	t.Add("CapturedKeeps", "int(11)").NotNullable()
-	t.Add("CapturedTowers", "int(11)").NotNullable()
-	t.Add("CapturedRelics", "int(11)").NotNullable()
-	t.Add("KillsDragon", "int(11)").NotNullable()
-	t.Add("DeathsPvP", "int(11)").NotNullable()
-	t.Add("KillsLegion", "int(11)").NotNullable()
-	t.Add("KillsEpicBoss", "int(11)").NotNullable()
-	t.Add("GainXP", "tinyint(1)").NotNullable()
-	t.Add("GainRP", "tinyint(1)").NotNullable()
-	t.Add("Autoloot", "tinyint(1)").NotNullable()
-	t.Add("LastFreeLeveled", "datetime")
-	t.Add("LastFreeLevel", "int(11)").NotNullable()
-	t.Add("GuildNote", "text")
-	t.Add("ShowXFireInfo", "tinyint(1)").NotNullable()
-	t.Add("NoHelp", "tinyint(1)").NotNullable()
-	t.Add("ShowGuildLogins", "tinyint(1)").NotNullable()
-	t.Add("Champion", "tinyint(1)").NotNullable()
-	t.Add("ChampionLevel", "int(11)").NotNullable()
-	t.Add("ChampionExperience", "bigint(20)").NotNullable()
-	t.Add("ML", "tinyint(3) unsigned").NotNullable()
-	t.Add("MLExperience", "bigint(20)").NotNullable()
-	t.Add("MLLevel", "int(11)").NotNullable()
-	t.Add("MLGranted", "tinyint(1)").NotNullable()
-	t.Add("RPFlag", "tinyint(1)").NotNullable()
-	t.Add("IgnoreStatistics", "tinyint(1)").NotNullable()
-	t.Add("NotDisplayedInHerald", "tinyint(3) unsigned").NotNullable()
-	t.Add("ActiveSaddleBags", "tinyint(3) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("DOLCharactersBackup_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DOLCharactersBackupXCustomParam")
-	t.AddWithIndex("DOLCharactersObjectId", "varchar(255)").NotNullable()
-	t.AddWithIndex("KeyName", "varchar(100)").NotNullable()
-	t.Add("Value", "varchar(255)")
-	t.AddPrimary("CustomParamID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("DOLCharacters")
-	t.Add("HasGravestone", "tinyint(1)").NotNullable()
-	t.Add("GravestoneRegion", "int(11)").NotNullable()
-	t.Add("Constitution", "int(11)").NotNullable()
-	t.Add("Dexterity", "int(11)").NotNullable()
-	t.Add("Strength", "int(11)").NotNullable()
-	t.Add("Quickness", "int(11)").NotNullable()
-	t.Add("Intelligence", "int(11)").NotNullable()
-	t.Add("Piety", "int(11)").NotNullable()
-	t.Add("Empathy", "int(11)").NotNullable()
-	t.Add("Charisma", "int(11)").NotNullable()
-	t.Add("BountyPoints", "bigint(20)").NotNullable()
-	t.Add("RealmPoints", "bigint(20)").NotNullable()
-	t.Add("RealmLevel", "int(11)").NotNullable()
-	t.Add("Experience", "bigint(20)").NotNullable()
-	t.Add("MaxEndurance", "int(11)").NotNullable()
-	t.Add("Health", "int(11)").NotNullable()
-	t.Add("Mana", "int(11)").NotNullable()
-	t.Add("Endurance", "int(11)").NotNullable()
-	t.Add("Concentration", "int(11)").NotNullable()
-	t.AddWithIndex("AccountName", "varchar(255)").NotNullable()
-	t.Add("AccountSlot", "int(11)").NotNullable()
-	t.Add("CreationDate", "datetime").NotNullable()
-	t.Add("LastPlayed", "datetime")
-	t.AddUnique("Name", "varchar(255)").NotNullable()
-	t.Add("LastName", "text")
-	t.AddWithIndex("GuildID", "varchar(255)")
-	t.Add("Gender", "int(11)").NotNullable()
-	t.Add("Race", "int(11)").NotNullable()
-	t.Add("Level", "int(11)").NotNullable()
-	t.Add("Class", "int(11)").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("CreationModel", "int(11)").NotNullable()
-	t.Add("CurrentModel", "int(11)").NotNullable()
-	t.Add("Region", "int(11)").NotNullable()
-	t.Add("ActiveWeaponSlot", "tinyint(3) unsigned").NotNullable()
-	t.Add("Xpos", "int(11)").NotNullable()
-	t.Add("Ypos", "int(11)").NotNullable()
-	t.Add("Zpos", "int(11)").NotNullable()
-	t.Add("BindXpos", "int(11)").NotNullable()
-	t.Add("BindYpos", "int(11)").NotNullable()
-	t.Add("BindZpos", "int(11)").NotNullable()
-	t.Add("BindRegion", "int(11)").NotNullable()
-	t.Add("BindHeading", "int(11)").NotNullable()
-	t.Add("BindHouseXpos", "int(11)").NotNullable()
-	t.Add("BindHouseYpos", "int(11)").NotNullable()
-	t.Add("BindHouseZpos", "int(11)").NotNullable()
-	t.Add("BindHouseRegion", "int(11)").NotNullable()
-	t.Add("BindHouseHeading", "int(11)").NotNullable()
-	t.Add("DeathCount", "tinyint(3) unsigned").NotNullable()
-	t.Add("ConLostAtDeath", "int(11)").NotNullable()
-	t.Add("Direction", "int(11)").NotNullable()
-	t.Add("MaxSpeed", "int(11)").NotNullable()
-	t.Add("Copper", "int(11)").NotNullable()
-	t.Add("Silver", "int(11)").NotNullable()
-	t.Add("Gold", "int(11)").NotNullable()
-	t.Add("Platinum", "int(11)").NotNullable()
-	t.Add("Mithril", "int(11)").NotNullable()
-	t.Add("SerializedCraftingSkills", "text")
-	t.Add("SerializedAbilities", "text")
-	t.Add("SerializedSpecs", "text")
-	t.Add("SerializedRealmAbilities", "text")
-	t.Add("DisabledSpells", "text")
-	t.Add("DisabledAbilities", "text")
-	t.Add("SerializedFriendsList", "text")
-	t.Add("SerializedIgnoreList", "text")
-	t.Add("IsCloakHoodUp", "tinyint(1)").NotNullable()
-	t.Add("IsCloakInvisible", "tinyint(1)").NotNullable()
-	t.Add("IsHelmInvisible", "tinyint(1)").NotNullable()
-	t.Add("SpellQueue", "tinyint(1)").NotNullable()
-	t.Add("IsLevelSecondStage", "tinyint(1)").NotNullable()
-	t.Add("FlagClassName", "tinyint(1)").NotNullable()
-	t.Add("Advisor", "tinyint(1)").NotNullable()
-	t.Add("GuildRank", "smallint(5) unsigned").NotNullable()
-	t.Add("PlayedTime", "bigint(20)").NotNullable()
-	t.Add("DeathTime", "bigint(20)").NotNullable()
-	t.Add("RespecAmountAllSkill", "int(11)").NotNullable()
-	t.Add("RespecAmountSingleSkill", "int(11)").NotNullable()
-	t.Add("RespecAmountRealmSkill", "int(11)").NotNullable()
-	t.Add("RespecAmountDOL", "int(11)").NotNullable()
-	t.Add("RespecAmountChampionSkill", "int(11)").NotNullable()
-	t.Add("IsLevelRespecUsed", "tinyint(1)").NotNullable()
-	t.Add("RespecBought", "int(11)").NotNullable()
-	t.Add("SafetyFlag", "tinyint(1)").NotNullable()
-	t.Add("CraftingPrimarySkill", "int(11)").NotNullable()
-	t.Add("CancelStyle", "tinyint(1)").NotNullable()
-	t.Add("IsAnonymous", "tinyint(1)").NotNullable()
-	t.Add("CustomisationStep", "tinyint(3) unsigned").NotNullable()
-	t.Add("EyeSize", "tinyint(3) unsigned").NotNullable()
-	t.Add("LipSize", "tinyint(3) unsigned").NotNullable()
-	t.Add("EyeColor", "tinyint(3) unsigned").NotNullable()
-	t.Add("HairColor", "tinyint(3) unsigned").NotNullable()
-	t.Add("FaceType", "tinyint(3) unsigned").NotNullable()
-	t.Add("HairStyle", "tinyint(3) unsigned").NotNullable()
-	t.Add("MoodType", "tinyint(3) unsigned").NotNullable()
-	t.Add("UsedLevelCommand", "tinyint(1)").NotNullable()
-	t.Add("CurrentTitleType", "text")
-	t.Add("KillsAlbionPlayers", "int(11)").NotNullable()
-	t.Add("KillsMidgardPlayers", "int(11)").NotNullable()
-	t.Add("KillsHiberniaPlayers", "int(11)").NotNullable()
-	t.Add("KillsAlbionDeathBlows", "int(11)").NotNullable()
-	t.Add("KillsMidgardDeathBlows", "int(11)").NotNullable()
-	t.Add("KillsHiberniaDeathBlows", "int(11)").NotNullable()
-	t.Add("KillsAlbionSolo", "int(11)").NotNullable()
-	t.Add("KillsMidgardSolo", "int(11)").NotNullable()
-	t.Add("KillsHiberniaSolo", "int(11)").NotNullable()
-	t.Add("CapturedKeeps", "int(11)").NotNullable()
-	t.Add("CapturedTowers", "int(11)").NotNullable()
-	t.Add("CapturedRelics", "int(11)").NotNullable()
-	t.Add("KillsDragon", "int(11)").NotNullable()
-	t.Add("DeathsPvP", "int(11)").NotNullable()
-	t.Add("KillsLegion", "int(11)").NotNullable()
-	t.Add("KillsEpicBoss", "int(11)").NotNullable()
-	t.Add("GainXP", "tinyint(1)").NotNullable()
-	t.Add("GainRP", "tinyint(1)").NotNullable()
-	t.Add("Autoloot", "tinyint(1)").NotNullable()
-	t.Add("LastFreeLeveled", "datetime")
-	t.Add("LastFreeLevel", "int(11)").NotNullable()
-	t.Add("GuildNote", "text")
-	t.Add("ShowXFireInfo", "tinyint(1)").NotNullable()
-	t.Add("NoHelp", "tinyint(1)").NotNullable()
-	t.Add("ShowGuildLogins", "tinyint(1)").NotNullable()
-	t.Add("Champion", "tinyint(1)").NotNullable()
-	t.Add("ChampionLevel", "int(11)").NotNullable()
-	t.Add("ChampionExperience", "bigint(20)").NotNullable()
-	t.Add("ML", "tinyint(3) unsigned").NotNullable()
-	t.Add("MLExperience", "bigint(20)").NotNullable()
-	t.Add("MLLevel", "int(11)").NotNullable()
-	t.Add("MLGranted", "tinyint(1)").NotNullable()
-	t.Add("RPFlag", "tinyint(1)").NotNullable()
-	t.Add("IgnoreStatistics", "tinyint(1)").NotNullable()
-	t.Add("NotDisplayedInHerald", "tinyint(3) unsigned").NotNullable()
-	t.Add("ActiveSaddleBags", "tinyint(3) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("DOLCharacters_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DOLCharactersXCustomParam")
-	t.AddWithIndex("DOLCharactersObjectId", "varchar(255)").NotNullable()
-	t.AddWithIndex("KeyName", "varchar(100)").NotNullable()
-	t.Add("Value", "varchar(255)")
-	t.AddPrimary("CustomParamID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("Door")
-	t.Add("Name", "text")
-	t.Add("Type", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.AddWithIndex("InternalID", "int(11)").NotNullable()
-	t.Add("Guild", "text")
-	t.Add("Level", "tinyint(3) unsigned").NotNullable()
-	t.Add("Realm", "tinyint(3) unsigned").NotNullable()
-	t.Add("Flags", "int(10) unsigned").NotNullable()
-	t.Add("Locked", "int(11)").NotNullable()
-	t.Add("Health", "int(11)").NotNullable()
-	t.Add("MaxHealth", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Door_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("DropTemplateXItemTemplate")
-	t.AddPrimary("ID", "bigint(20)").NotNullable()
-	t.AddWithIndex("TemplateName", "varchar(255)").NotNullable()
-	t.Add("ItemTemplateID", "text").NotNullable()
-	t.Add("Chance", "int(11)").NotNullable()
-	t.Add("Count", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("FactionAggroLevel")
-	t.AddWithIndex("CharacterID", "varchar(100)").NotNullable()
-	t.Add("FactionID", "int(11)").NotNullable()
-	t.Add("AggroLevel", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("FactionAggroLevel_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Faction")
-	t.AddPrimary("ID", "int(11)").NotNullable()
-	t.Add("Name", "text")
-	t.Add("BaseAggroLevel", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Faction_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("GuildAlliance")
-	t.Add("AllianceName", "text")
-	t.Add("Motd", "text")
-	t.AddUnique("LeaderGuildID", "varchar(255)")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("GuildAlliance_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("GuildRank")
-	t.AddWithIndex("GuildID", "varchar(255)")
-	t.Add("Title", "text")
-	t.Add("RankLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("Alli", "tinyint(1)").NotNullable()
-	t.Add("Emblem", "tinyint(1)").NotNullable()
-	t.Add("Buff", "tinyint(1)").NotNullable()
-	t.Add("GcHear", "tinyint(1)").NotNullable()
-	t.Add("GcSpeak", "tinyint(1)").NotNullable()
-	t.Add("OcHear", "tinyint(1)").NotNullable()
-	t.Add("OcSpeak", "tinyint(1)").NotNullable()
-	t.Add("AcHear", "tinyint(1)").NotNullable()
-	t.Add("AcSpeak", "tinyint(1)").NotNullable()
-	t.Add("Invite", "tinyint(1)").NotNullable()
-	t.Add("Promote", "tinyint(1)").NotNullable()
-	t.Add("Remove", "tinyint(1)").NotNullable()
-	t.Add("View", "tinyint(1)").NotNullable()
-	t.Add("Claim", "tinyint(1)").NotNullable()
-	t.Add("Upgrade", "tinyint(1)").NotNullable()
-	t.Add("Release", "tinyint(1)").NotNullable()
-	t.Add("Dues", "tinyint(1)").NotNullable()
-	t.Add("Withdraw", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("GuildRank_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Guild")
-	t.AddUnique("GuildID", "varchar(255)")
-	t.AddWithIndex("GuildName", "varchar(255)")
-	t.Add("Realm", "tinyint(3) unsigned").NotNullable()
-	t.Add("GuildBanner", "tinyint(1)").NotNullable()
-	t.Add("GuildBannerLostTime", "datetime").NotNullable()
-	t.Add("Motd", "text")
-	t.Add("oMotd", "text")
-	t.AddWithIndex("AllianceID", "varchar(255)")
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("RealmPoints", "bigint(20)").NotNullable()
-	t.Add("BountyPoints", "bigint(20)").NotNullable()
-	t.Add("Webpage", "text")
-	t.Add("Email", "text")
-	t.Add("Dues", "tinyint(1)").NotNullable()
-	t.Add("Bank", "double").NotNullable()
-	t.Add("DuesPercent", "bigint(20)").NotNullable()
-	t.Add("HaveGuildHouse", "tinyint(1)").NotNullable()
-	t.Add("GuildHouseNumber", "int(11)").NotNullable()
-	t.Add("GuildLevel", "bigint(20)").NotNullable()
-	t.Add("BonusType", "tinyint(3) unsigned").NotNullable()
-	t.Add("BonusStartTime", "datetime").NotNullable()
-	t.Add("MeritPoints", "bigint(20)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Guild_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("HouseConsignmentMerchant")
-	t.AddPrimary("ID", "bigint(20)").NotNullable()
-	t.AddWithIndex("OwnerID", "varchar(128)").NotNullable()
-	t.AddWithIndex("HouseNumber", "int(11)").NotNullable()
-	t.Add("Money", "bigint(20)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("househookpointitem")
-	t.AddPrimary("ID", "bigint(20)").NotNullable()
-	t.AddWithIndex("HouseNumber", "int(11)").NotNullable()
-	t.Add("HookpointID", "int(10) unsigned").NotNullable()
-	t.Add("Heading", "smallint(5) unsigned").NotNullable()
-	t.Add("ItemTemplateID", "text")
-	t.Add("Index", "tinyint(3) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 10019
-	allTables = append(allTables, *t)
-
-	t = NewTable("househookpointoffset")
-	t.AddPrimary("ID", "bigint(20)").NotNullable()
-	t.Add("HouseModel", "int(11)").NotNullable()
-	t.Add("HookpointID", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 144
-	allTables = append(allTables, *t)
-
-	t = NewTable("InstanceXElement")
-	t.AddWithIndex("InstanceID", "varchar(255)").NotNullable()
-	t.Add("ClassType", "text")
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Heading", "smallint(5) unsigned").NotNullable()
-	t.Add("NPCTemplate", "varchar(255)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("InstanceXElement_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Inventory")
-	t.AddWithIndex("OwnerID", "varchar(255)").NotNullable()
-	t.Add("OwnerLot", "smallint(5) unsigned").NotNullable()
-	t.AddWithIndex("ITemplate_Id", "varchar(255)")
-	t.AddWithIndex("UTemplate_Id", "varchar(255)")
-	t.Add("IsCrafted", "tinyint(1)").NotNullable()
-	t.Add("Creator", "text")
-	t.AddWithIndex("SlotPosition", "int(11)").NotNullable()
-	t.Add("Count", "int(11)").NotNullable()
-	t.Add("SellPrice", "int(11)").NotNullable()
-	t.Add("Experience", "bigint(20)").NotNullable()
-	t.Add("Color", "int(11)").NotNullable()
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("Extension", "tinyint(3) unsigned").NotNullable()
-	t.Add("Condition", "int(11)").NotNullable()
-	t.Add("Durability", "int(11)").NotNullable()
-	t.Add("PoisonSpellID", "int(11)").NotNullable()
-	t.Add("PoisonMaxCharges", "int(11)").NotNullable()
-	t.Add("PoisonCharges", "int(11)").NotNullable()
-	t.Add("Charges", "int(11)").NotNullable()
-	t.Add("Charges1", "int(11)").NotNullable()
-	t.Add("Cooldown", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Inventory_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("ItemTemplate")
-	t.AddPrimary("Id_nb", "varchar(255)").NotNullable()
-	t.Add("TranslationId", "text")
-	t.Add("Name", "text").NotNullable()
-	t.Add("ExamineArticle", "text")
-	t.Add("MessageArticle", "text")
-	t.Add("Level", "int(11)").NotNullable()
-	t.Add("Durability", "int(11)").NotNullable()
-	t.Add("MaxDurability", "int(11)").NotNullable()
-	t.Add("IsNotLosingDur", "tinyint(1)").NotNullable()
-	t.Add("Condition", "int(11)").NotNullable()
-	t.Add("MaxCondition", "int(11)").NotNullable()
-	t.Add("Quality", "int(11)").NotNullable()
-	t.Add("DPS_AF", "int(11)").NotNullable()
-	t.Add("SPD_ABS", "int(11)").NotNullable()
-	t.Add("Hand", "int(11)").NotNullable()
-	t.Add("Type_Damage", "int(11)").NotNullable()
-	t.Add("Object_Type", "int(11)").NotNullable()
-	t.Add("Item_Type", "int(11)").NotNullable()
-	t.Add("Color", "int(11)").NotNullable()
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("Effect", "int(11)").NotNullable()
-	t.Add("Weight", "int(11)").NotNullable()
-	t.Add("Model", "int(11)").NotNullable()
-	t.Add("Extension", "tinyint(3) unsigned").NotNullable()
-	t.Add("Bonus", "int(11)").NotNullable()
-	t.Add("Bonus1", "int(11)").NotNullable()
-	t.Add("Bonus2", "int(11)").NotNullable()
-	t.Add("Bonus3", "int(11)").NotNullable()
-	t.Add("Bonus4", "int(11)").NotNullable()
-	t.Add("Bonus5", "int(11)").NotNullable()
-	t.Add("Bonus6", "int(11)").NotNullable()
-	t.Add("Bonus7", "int(11)").NotNullable()
-	t.Add("Bonus8", "int(11)").NotNullable()
-	t.Add("Bonus9", "int(11)").NotNullable()
-	t.Add("Bonus10", "int(11)").NotNullable()
-	t.Add("ExtraBonus", "int(11)").NotNullable()
-	t.Add("Bonus1Type", "int(11)").NotNullable()
-	t.Add("Bonus2Type", "int(11)").NotNullable()
-	t.Add("Bonus3Type", "int(11)").NotNullable()
-	t.Add("Bonus4Type", "int(11)").NotNullable()
-	t.Add("Bonus5Type", "int(11)").NotNullable()
-	t.Add("Bonus6Type", "int(11)").NotNullable()
-	t.Add("Bonus7Type", "int(11)").NotNullable()
-	t.Add("Bonus8Type", "int(11)").NotNullable()
-	t.Add("Bonus9Type", "int(11)").NotNullable()
-	t.Add("Bonus10Type", "int(11)").NotNullable()
-	t.Add("ExtraBonusType", "int(11)").NotNullable()
-	t.Add("CanUseEvery", "int(11)").NotNullable()
-	t.Add("SpellID", "int(11)").NotNullable()
-	t.Add("Charges", "int(11)").NotNullable()
-	t.Add("MaxCharges", "int(11)").NotNullable()
-	t.Add("SpellID1", "int(11)").NotNullable()
-	t.Add("Charges1", "int(11)").NotNullable()
-	t.Add("MaxCharges1", "int(11)").NotNullable()
-	t.Add("ProcChance", "tinyint(3) unsigned").NotNullable()
-	t.Add("ProcSpellID", "int(11)").NotNullable()
-	t.Add("ProcSpellID1", "int(11)").NotNullable()
-	t.Add("PoisonSpellID", "int(11)").NotNullable()
-	t.Add("PoisonCharges", "int(11)").NotNullable()
-	t.Add("PoisonMaxCharges", "int(11)").NotNullable()
-	t.Add("IsPickable", "tinyint(1)").NotNullable()
-	t.Add("IsDropable", "tinyint(1)").NotNullable()
-	t.Add("CanDropAsLoot", "tinyint(1)").NotNullable()
-	t.Add("IsTradable", "tinyint(1)").NotNullable()
-	t.Add("Price", "bigint(20)").NotNullable()
-	t.Add("MaxCount", "int(11)").NotNullable()
-	t.Add("IsIndestructible", "tinyint(1)").NotNullable()
-	t.Add("PackSize", "int(11)").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("AllowedClasses", "text").NotNullable()
-	t.Add("Flags", "int(11)").NotNullable()
-	t.Add("BonusLevel", "int(11)").NotNullable()
-	t.Add("LevelRequirement", "int(11)").NotNullable()
-	t.Add("PackageID", "text")
-	t.Add("Description", "text")
-	t.Add("ClassType", "varchar(255)")
-	t.Add("SalvageYieldID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("ItemTemplate_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("ItemUnique")
-	t.AddPrimary("Id_nb", "varchar(255)").NotNullable()
-	t.Add("TranslationId", "text")
-	t.Add("Name", "text").NotNullable()
-	t.Add("ExamineArticle", "text")
-	t.Add("MessageArticle", "text")
-	t.Add("Level", "int(11)").NotNullable()
-	t.Add("Durability", "int(11)").NotNullable()
-	t.Add("MaxDurability", "int(11)").NotNullable()
-	t.Add("IsNotLosingDur", "tinyint(1)").NotNullable()
-	t.Add("Condition", "int(11)").NotNullable()
-	t.Add("MaxCondition", "int(11)").NotNullable()
-	t.Add("Quality", "int(11)").NotNullable()
-	t.Add("DPS_AF", "int(11)").NotNullable()
-	t.Add("SPD_ABS", "int(11)").NotNullable()
-	t.Add("Hand", "int(11)").NotNullable()
-	t.Add("Type_Damage", "int(11)").NotNullable()
-	t.Add("Object_Type", "int(11)").NotNullable()
-	t.Add("Item_Type", "int(11)").NotNullable()
-	t.Add("Color", "int(11)").NotNullable()
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("Effect", "int(11)").NotNullable()
-	t.Add("Weight", "int(11)").NotNullable()
-	t.Add("Model", "int(11)").NotNullable()
-	t.Add("Extension", "tinyint(3) unsigned").NotNullable()
-	t.Add("Bonus", "int(11)").NotNullable()
-	t.Add("Bonus1", "int(11)").NotNullable()
-	t.Add("Bonus2", "int(11)").NotNullable()
-	t.Add("Bonus3", "int(11)").NotNullable()
-	t.Add("Bonus4", "int(11)").NotNullable()
-	t.Add("Bonus5", "int(11)").NotNullable()
-	t.Add("Bonus6", "int(11)").NotNullable()
-	t.Add("Bonus7", "int(11)").NotNullable()
-	t.Add("Bonus8", "int(11)").NotNullable()
-	t.Add("Bonus9", "int(11)").NotNullable()
-	t.Add("Bonus10", "int(11)").NotNullable()
-	t.Add("ExtraBonus", "int(11)").NotNullable()
-	t.Add("Bonus1Type", "int(11)").NotNullable()
-	t.Add("Bonus2Type", "int(11)").NotNullable()
-	t.Add("Bonus3Type", "int(11)").NotNullable()
-	t.Add("Bonus4Type", "int(11)").NotNullable()
-	t.Add("Bonus5Type", "int(11)").NotNullable()
-	t.Add("Bonus6Type", "int(11)").NotNullable()
-	t.Add("Bonus7Type", "int(11)").NotNullable()
-	t.Add("Bonus8Type", "int(11)").NotNullable()
-	t.Add("Bonus9Type", "int(11)").NotNullable()
-	t.Add("Bonus10Type", "int(11)").NotNullable()
-	t.Add("ExtraBonusType", "int(11)").NotNullable()
-	t.Add("CanUseEvery", "int(11)").NotNullable()
-	t.Add("SpellID", "int(11)").NotNullable()
-	t.Add("Charges", "int(11)").NotNullable()
-	t.Add("MaxCharges", "int(11)").NotNullable()
-	t.Add("SpellID1", "int(11)").NotNullable()
-	t.Add("Charges1", "int(11)").NotNullable()
-	t.Add("MaxCharges1", "int(11)").NotNullable()
-	t.Add("ProcChance", "tinyint(3) unsigned").NotNullable()
-	t.Add("ProcSpellID", "int(11)").NotNullable()
-	t.Add("ProcSpellID1", "int(11)").NotNullable()
-	t.Add("PoisonSpellID", "int(11)").NotNullable()
-	t.Add("PoisonCharges", "int(11)").NotNullable()
-	t.Add("PoisonMaxCharges", "int(11)").NotNullable()
-	t.Add("IsPickable", "tinyint(1)").NotNullable()
-	t.Add("IsDropable", "tinyint(1)").NotNullable()
-	t.Add("CanDropAsLoot", "tinyint(1)").NotNullable()
-	t.Add("IsTradable", "tinyint(1)").NotNullable()
-	t.Add("Price", "bigint(20)").NotNullable()
-	t.Add("MaxCount", "int(11)").NotNullable()
-	t.Add("IsIndestructible", "tinyint(1)").NotNullable()
-	t.Add("PackSize", "int(11)").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("AllowedClasses", "text").NotNullable()
-	t.Add("Flags", "int(11)").NotNullable()
-	t.Add("BonusLevel", "int(11)").NotNullable()
-	t.Add("LevelRequirement", "int(11)").NotNullable()
-	t.Add("PackageID", "text")
-	t.Add("Description", "text")
-	t.Add("ClassType", "varchar(255)")
-	t.Add("SalvageYieldID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("ItemUnique_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("KeepCaptureLog")
-	t.AddPrimary("ID", "bigint(20)").NotNullable()
-	t.Add("DateTaken", "datetime").NotNullable()
-	t.Add("KeepName", "text").NotNullable()
-	t.Add("KeepType", "text").NotNullable()
-	t.Add("NumEnemies", "int(11)").NotNullable()
-	t.Add("CombatTime", "int(11)").NotNullable()
-	t.Add("RPReward", "int(11)").NotNullable()
-	t.Add("BPReward", "int(11)").NotNullable()
-	t.Add("XPReward", "bigint(20)").NotNullable()
-	t.Add("MoneyReward", "bigint(20)").NotNullable()
-	t.Add("CapturedBy", "text").NotNullable()
-	t.Add("RPGainerList", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("KeepComponent")
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.Add("Health", "int(11)").NotNullable()
-	t.Add("Skin", "int(11)").NotNullable()
-	t.AddWithIndex("KeepID", "int(11)").NotNullable()
-	t.AddWithIndex("ID", "int(11)").NotNullable()
-	t.Add("CreateInfo", "varchar(255)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("KeepComponent_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("KeepHookPointItem")
-	t.AddWithIndex("KeepID", "int(11)").NotNullable()
-	t.AddWithIndex("ComponentID", "int(11)").NotNullable()
-	t.AddWithIndex("HookPointID", "int(11)").NotNullable()
-	t.Add("ClassType", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("KeepHookPointItem_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("KeepHookPoint")
-	t.AddWithIndex("HookPointID", "int(11)").NotNullable()
-	t.Add("KeepComponentSkinID", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.AddWithIndex("Height", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("KeepHookPoint_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("KeepPosition")
-	t.AddWithIndex("ComponentSkin", "int(11)").NotNullable()
-	t.Add("ComponentRotation", "int(11)").NotNullable()
-	t.AddWithIndex("TemplateID", "varchar(255)").NotNullable()
-	t.AddWithIndex("Height", "int(11)").NotNullable()
-	t.Add("XOff", "int(11)").NotNullable()
-	t.Add("YOff", "int(11)").NotNullable()
-	t.Add("ZOff", "int(11)").NotNullable()
-	t.Add("HOff", "int(11)").NotNullable()
-	t.AddWithIndex("ClassType", "varchar(255)")
-	t.Add("TemplateType", "int(11)").NotNullable()
-	t.Add("KeepType", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("KeepPosition_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Keep")
-	t.AddPrimary("KeepID", "int(11)").NotNullable()
-	t.Add("Name", "text").NotNullable()
-	t.Add("Region", "smallint(5) unsigned").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Heading", "smallint(5) unsigned").NotNullable()
-	t.Add("Realm", "tinyint(3) unsigned").NotNullable()
-	t.Add("Level", "tinyint(3) unsigned").NotNullable()
-	t.Add("ClaimedGuildName", "text")
-	t.Add("AlbionDifficultyLevel", "int(11)").NotNullable()
-	t.Add("MidgardDifficultyLevel", "int(11)").NotNullable()
-	t.Add("HiberniaDifficultyLevel", "int(11)").NotNullable()
-	t.Add("OriginalRealm", "int(11)").NotNullable()
-	t.Add("KeepType", "int(11)").NotNullable()
-	t.Add("BaseLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("SkinType", "tinyint(3) unsigned").NotNullable()
-	t.Add("CreateInfo", "varchar(255)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Keep_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("LanguageArea")
-	t.Add("Description", "text").NotNullable()
-	t.Add("ScreenDescription", "text").NotNullable()
-	t.AddWithIndex("TranslationId", "varchar(255)").NotNullable()
-	t.AddWithIndex("Language", "varchar(255)").NotNullable()
-	t.Add("Tag", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LanguageArea_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LanguageGameObject")
-	t.Add("Name", "text")
-	t.Add("ExamineArticle", "text")
-	t.AddWithIndex("TranslationId", "varchar(255)").NotNullable()
-	t.AddWithIndex("Language", "varchar(255)").NotNullable()
-	t.Add("Tag", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LanguageGameObject_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LanguageNPC")
-	t.Add("Name", "text")
-	t.Add("Suffix", "text")
-	t.Add("GuildName", "text")
-	t.Add("ExamineArticle", "text")
-	t.Add("MessageArticle", "text")
-	t.AddWithIndex("TranslationId", "varchar(255)").NotNullable()
-	t.AddWithIndex("Language", "varchar(255)").NotNullable()
-	t.Add("Tag", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LanguageNPC_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LanguageSystem")
-	t.Add("Text", "text").NotNullable()
-	t.AddWithIndex("TranslationId", "varchar(255)").NotNullable()
-	t.AddWithIndex("Language", "varchar(255)").NotNullable()
-	t.Add("Tag", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LanguageSystem_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LanguageZone")
-	t.Add("Description", "text").NotNullable()
-	t.Add("ScreenDescription", "text").NotNullable()
-	t.AddWithIndex("TranslationId", "varchar(255)").NotNullable()
-	t.AddWithIndex("Language", "varchar(255)").NotNullable()
-	t.Add("Tag", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LanguageZone_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LineXSpell")
-	t.AddWithIndex("LineName", "varchar(255)").NotNullable()
-	t.Add("SpellID", "int(11)").NotNullable()
-	t.Add("Level", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LineXSpell_ID", "varchar(255)").NotNullable()
-	t.Add("PackageID", "text")
-	allTables = append(allTables, *t)
-
-	t = NewTable("LinkedFaction")
-	t.Add("FactionID", "int(11)").NotNullable()
-	t.Add("LinkedFactionID", "int(11)").NotNullable()
-	t.Add("IsFriend", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LinkedFaction_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LootGenerator")
-	t.Add("MobName", "text")
-	t.Add("MobGuild", "text")
-	t.Add("MobFaction", "text")
-	t.Add("RegionID", "smallint(5) unsigned").NotNullable()
-	t.Add("LootGeneratorClass", "text").NotNullable()
-	t.Add("ExclusivePriority", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LootGenerator_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LootOTD")
-	t.AddWithIndex("MobName", "varchar(100)").NotNullable()
-	t.AddWithIndex("ItemTemplateID", "varchar(100)").NotNullable()
-	t.Add("MinLevel", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LootOTD_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("LootTemplate")
-	t.AddWithIndex("TemplateName", "varchar(255)").NotNullable()
-	t.Add("ItemTemplateID", "text").NotNullable()
-	t.Add("Chance", "int(11)").NotNullable()
-	t.Add("Count", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("LootTemplate_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("MerchantItem")
-	t.AddWithIndex("ItemListID", "varchar(255)").NotNullable()
-	t.Add("ItemTemplateID", "text").NotNullable()
-	t.AddWithIndex("PageNumber", "int(11)").NotNullable()
-	t.AddWithIndex("SlotPosition", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("MerchantItem_ID", "varchar(255)").NotNullable()
-	t.Add("Price", "bigint(20)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Minotaurrelic")
-	t.Add("relicSpell", "int(11)").NotNullable()
-	t.Add("SpawnLocked", "tinyint(1)").NotNullable()
-	t.Add("ProtectorClassType", "text").NotNullable()
-	t.Add("relicTarget", "text").NotNullable()
-	t.Add("Name", "text").NotNullable()
-	t.Add("Model", "smallint(5) unsigned").NotNullable()
-	t.Add("SpawnX", "int(11)").NotNullable()
-	t.Add("SpawnY", "int(11)").NotNullable()
-	t.Add("SpawnZ", "int(11)").NotNullable()
-	t.Add("SpawnHeading", "int(11)").NotNullable()
-	t.Add("SpawnRegion", "int(11)").NotNullable()
-	t.Add("Effect", "int(11)").NotNullable()
-	t.AddPrimary("RelicID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Minotaurrelic_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("MobDropTemplate")
-	t.AddPrimary("ID", "bigint(20)").NotNullable()
-	t.AddWithIndex("MobName", "varchar(255)").NotNullable()
-	t.AddWithIndex("LootTemplateName", "varchar(255)").NotNullable()
-	t.Add("DropCount", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("Mob")
-	t.Add("ClassType", "text")
-	t.Add("TranslationId", "text")
-	t.AddWithIndex("Name", "varchar(255)").NotNullable()
-	t.Add("Suffix", "text")
-	t.Add("Guild", "text")
-	t.Add("ExamineArticle", "text")
-	t.Add("MessageArticle", "text")
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Speed", "int(11)").NotNullable()
-	t.Add("Heading", "smallint(5) unsigned").NotNullable()
-	t.Add("Region", "smallint(5) unsigned").NotNullable()
-	t.Add("Model", "smallint(5) unsigned").NotNullable()
-	t.Add("Size", "tinyint(3) unsigned").NotNullable()
-	t.Add("Strength", "smallint(6)").NotNullable()
-	t.Add("Constitution", "smallint(6)").NotNullable()
-	t.Add("Dexterity", "smallint(6)").NotNullable()
-	t.Add("Quickness", "smallint(6)").NotNullable()
-	t.Add("Intelligence", "smallint(6)").NotNullable()
-	t.Add("Piety", "smallint(6)").NotNullable()
-	t.Add("Empathy", "smallint(6)").NotNullable()
-	t.Add("Charisma", "smallint(6)").NotNullable()
-	t.Add("Level", "tinyint(3) unsigned").NotNullable()
-	t.Add("Realm", "tinyint(3) unsigned").NotNullable()
-	t.Add("EquipmentTemplateID", "text")
-	t.Add("ItemsListTemplateID", "text")
-	t.Add("NPCTemplateID", "int(11)").NotNullable()
-	t.Add("Race", "int(11)").NotNullable()
-	t.Add("Flags", "int(10) unsigned").NotNullable()
-	t.Add("AggroLevel", "int(11)").NotNullable()
-	t.Add("AggroRange", "int(11)").NotNullable()
-	t.Add("MeleeDamageType", "int(11)").NotNullable()
-	t.Add("RespawnInterval", "int(11)").NotNullable()
-	t.Add("FactionID", "int(11)").NotNullable()
-	t.Add("BodyType", "int(11)").NotNullable()
-	t.AddWithIndex("HouseNumber", "int(11)").NotNullable()
-	t.Add("Brain", "text")
-	t.Add("PathID", "text")
-	t.Add("MaxDistance", "int(11)").NotNullable()
-	t.Add("OwnerID", "varchar(255)").NotNullable()
-	t.Add("RoamingRange", "int(11)").NotNullable()
-	t.Add("IsCloakHoodUp", "tinyint(1)").NotNullable()
-	t.Add("Gender", "tinyint(3) unsigned").NotNullable()
-	t.Add("PackageID", "text")
-	t.Add("VisibleWeaponSlots", "tinyint(3) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Mob_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("MobXAmbientBehaviour")
-	t.AddWithIndex("Source", "varchar(255)").NotNullable()
-	t.Add("Trigger", "text").NotNullable()
-	t.Add("Emote", "smallint(5) unsigned").NotNullable()
-	t.Add("Text", "text").NotNullable()
-	t.Add("Chance", "smallint(5) unsigned").NotNullable()
-	t.Add("Voice", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("MobXAmbientBehaviour_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("MobXLootTemplate")
-	t.AddWithIndex("MobName", "varchar(255)").NotNullable()
-	t.AddWithIndex("LootTemplateName", "varchar(255)").NotNullable()
-	t.Add("DropCount", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("MobXLootTemplate_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("News")
-	t.Add("CreationDate", "datetime").NotNullable()
-	t.Add("Type", "tinyint(3) unsigned").NotNullable()
-	t.Add("Realm", "tinyint(3) unsigned").NotNullable()
-	t.Add("Text", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("News_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("NPCEquipment")
-	t.AddWithIndex("TemplateID", "varchar(255)").NotNullable()
-	t.Add("Slot", "int(11)").NotNullable()
-	t.Add("Model", "int(11)").NotNullable()
-	t.Add("Color", "int(11)").NotNullable()
-	t.Add("Effect", "int(11)").NotNullable()
-	t.Add("Extension", "int(11)").NotNullable()
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("NPCEquipment_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("NpcTemplate")
-	t.AddWithIndex("TemplateId", "int(11)").NotNullable()
-	t.Add("TranslationId", "text")
-	t.Add("Name", "text").NotNullable()
-	t.Add("Suffix", "text")
-	t.Add("ClassType", "text").NotNullable()
-	t.Add("GuildName", "text")
-	t.Add("ExamineArticle", "text")
-	t.Add("MessageArticle", "text")
-	t.Add("Model", "text").NotNullable()
-	t.Add("Gender", "smallint(5) unsigned").NotNullable()
-	t.Add("Size", "text").NotNullable()
-	t.Add("Level", "text").NotNullable()
-	t.Add("MaxSpeed", "smallint(6)").NotNullable()
-	t.Add("EquipmentTemplateID", "text")
-	t.Add("ItemsListTemplateID", "text")
-	t.Add("Flags", "smallint(5) unsigned").NotNullable()
-	t.Add("MeleeDamageType", "tinyint(3) unsigned").NotNullable()
-	t.Add("ParryChance", "tinyint(3) unsigned").NotNullable()
-	t.Add("EvadeChance", "tinyint(3) unsigned").NotNullable()
-	t.Add("BlockChance", "tinyint(3) unsigned").NotNullable()
-	t.Add("LeftHandSwingChance", "tinyint(3) unsigned").NotNullable()
-	t.Add("Spells", "text")
-	t.Add("Styles", "text")
-	t.Add("Strength", "smallint(6)").NotNullable()
-	t.Add("Constitution", "smallint(6)").NotNullable()
-	t.Add("Dexterity", "smallint(6)").NotNullable()
-	t.Add("Quickness", "smallint(6)").NotNullable()
-	t.Add("Intelligence", "smallint(6)").NotNullable()
-	t.Add("Piety", "smallint(6)").NotNullable()
-	t.Add("Charisma", "smallint(6)").NotNullable()
-	t.Add("Empathy", "smallint(6)").NotNullable()
-	t.Add("Abilities", "text")
-	t.Add("AggroLevel", "tinyint(3) unsigned").NotNullable()
-	t.Add("AggroRange", "int(11)").NotNullable()
-	t.Add("Race", "int(11)").NotNullable()
-	t.Add("BodyType", "int(11)").NotNullable()
-	t.Add("MaxDistance", "int(11)").NotNullable()
-	t.Add("TetherRange", "int(11)").NotNullable()
-	t.Add("VisibleWeaponSlots", "tinyint(3) unsigned").NotNullable()
-	t.Add("ReplaceMobValues", "tinyint(1)").NotNullable()
-	t.Add("PackageID", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("NpcTemplate_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("PathPoints")
-	t.AddWithIndex("PathID", "varchar(255)").NotNullable()
-	t.Add("Step", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("MaxSpeed", "int(11)").NotNullable()
-	t.Add("WaitTime", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("PathPoints_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Path")
-	t.AddUnique("PathID", "varchar(255)").NotNullable()
-	t.Add("PathType", "int(11)").NotNullable()
-	t.Add("RegionID", "smallint(5) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Path_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("PlayerBoats")
-	t.Add("BoatID", "text").NotNullable()
-	t.Add("BoatOwner", "text").NotNullable()
-	t.AddPrimary("BoatName", "varchar(255)").NotNullable()
-	t.Add("BoatModel", "smallint(5) unsigned").NotNullable()
-	t.Add("BoatMaxSpeedBase", "smallint(6)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("PlayerBoats_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("PlayerInfo")
-	t.Add("Name", "text")
-	t.Add("LastName", "text")
-	t.Add("Guild", "text")
-	t.Add("Race", "text")
-	t.Add("Class", "text")
-	t.Add("Alive", "text")
-	t.Add("Realm", "text")
-	t.Add("Region", "text")
-	t.Add("Level", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("PlayerInfo_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("PlayerXEffect")
-	t.Add("IsHandler", "tinyint(1)").NotNullable()
-	t.Add("Var6", "int(11)").NotNullable()
-	t.Add("Var5", "int(11)").NotNullable()
-	t.Add("Var4", "int(11)").NotNullable()
-	t.Add("Var3", "int(11)").NotNullable()
-	t.Add("Var2", "int(11)").NotNullable()
-	t.Add("Var1", "int(11)").NotNullable()
-	t.Add("Duration", "int(11)").NotNullable()
-	t.Add("EffectType", "text")
-	t.Add("SpellLine", "text")
-	t.AddWithIndex("ChardID", "varchar(255)")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("PlayerXEffect_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("PvPKillsLog")
-	t.AddPrimary("ID", "bigint(20)").NotNullable()
-	t.Add("DateKilled", "datetime").NotNullable()
-	t.Add("KilledName", "text").NotNullable()
-	t.Add("KillerName", "text").NotNullable()
-	t.Add("KillerIP", "text").NotNullable()
-	t.Add("KilledIP", "text").NotNullable()
-	t.Add("KilledRealm", "text").NotNullable()
-	t.Add("KillerRealm", "text").NotNullable()
-	t.Add("RPReward", "int(11)").NotNullable()
-	t.Add("SameIP", "tinyint(3) unsigned").NotNullable()
-	t.Add("RegionName", "text")
-	t.Add("IsInstance", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("Quest")
-	t.Add("Name", "text").NotNullable()
-	t.Add("Step", "int(11)").NotNullable()
-	t.AddWithIndex("Character_ID", "varchar(255)").NotNullable()
-	t.Add("CustomPropertiesString", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Quest_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Race")
-	t.AddUnique("ID", "int(11)").NotNullable()
-	t.AddUnique("Name", "varchar(255)").NotNullable()
-	t.Add("ResistBody", "tinyint(3)").NotNullable()
-	t.Add("ResistCold", "tinyint(3)").NotNullable()
-	t.Add("ResistCrush", "tinyint(3)").NotNullable()
-	t.Add("ResistEnergy", "tinyint(3)").NotNullable()
-	t.Add("ResistHeat", "tinyint(3)").NotNullable()
-	t.Add("ResistMatter", "tinyint(3)").NotNullable()
-	t.Add("ResistNatural", "tinyint(3)").NotNullable()
-	t.Add("ResistSlash", "tinyint(3)").NotNullable()
-	t.Add("ResistSpirit", "tinyint(3)").NotNullable()
-	t.Add("ResistThrust", "tinyint(3)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Race_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Regions")
-	t.AddPrimary("RegionID", "smallint(5) unsigned").NotNullable()
-	t.Add("Name", "text").NotNullable()
-	t.Add("Description", "text").NotNullable()
-	t.Add("IP", "text").NotNullable()
-	t.Add("Port", "smallint(5) unsigned").NotNullable()
-	t.Add("Expansion", "int(11)").NotNullable()
-	t.Add("HousingEnabled", "tinyint(1)").NotNullable()
-	t.Add("DivingEnabled", "tinyint(1)").NotNullable()
-	t.Add("WaterLevel", "int(11)").NotNullable()
-	t.Add("ClassType", "varchar(200)").NotNullable()
-	t.Add("IsFrontier", "tinyint(1)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Regions_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("Relic")
-	t.AddPrimary("RelicID", "int(11)").NotNullable()
-	t.Add("Region", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("OriginalRealm", "int(11)").NotNullable()
-	t.Add("LastRealm", "int(11)").NotNullable()
-	t.Add("relicType", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Relic_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("Salvage")
-	t.AddWithIndex("ObjectType", "int(11)").NotNullable()
-	t.AddWithIndex("SalvageLevel", "int(11)").NotNullable()
-	t.Add("Id_nb", "text").NotNullable()
-	t.AddWithIndex("Realm", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Salvage_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("SalvageYield")
-	t.AddPrimary("ID", "int(11)").NotNullable()
-	t.AddWithIndex("ObjectType", "int(11)").NotNullable()
-	t.AddWithIndex("SalvageLevel", "int(11)").NotNullable()
-	t.Add("MaterialId_nb", "text").NotNullable()
-	t.Add("Count", "int(11)").NotNullable()
-	t.AddWithIndex("Realm", "int(11)").NotNullable()
-	t.Add("PackageID", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 366
-	allTables = append(allTables, *t)
-
-	t = NewTable("ServerInfo")
-	t.Add("Time", "text")
-	t.Add("ServerName", "text")
-	t.Add("AAC", "text")
-	t.Add("ServerType", "text")
-	t.Add("ServerStatus", "text")
-	t.Add("NumClients", "int(11)").NotNullable()
-	t.Add("NumAccounts", "int(11)").NotNullable()
-	t.Add("NumMobs", "int(11)").NotNullable()
-	t.Add("NumInventoryItems", "int(11)").NotNullable()
-	t.Add("NumPlayerChars", "int(11)").NotNullable()
-	t.Add("NumMerchantItems", "int(11)").NotNullable()
-	t.Add("NumItemTemplates", "int(11)").NotNullable()
-	t.Add("NumWorldObjects", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("ServerInfo_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("serverproperty_category")
-	t.Add("BaseCategory", "text").NotNullable()
-	t.Add("ParentCategory", "text")
-	t.Add("DisplayName", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("serverproperty_category_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("ServerProperty")
-	t.Add("Category", "text").NotNullable()
-	t.AddPrimary("Key", "varchar(255)").NotNullable()
-	t.Add("Description", "text").NotNullable()
-	t.Add("DefaultValue", "text").NotNullable()
-	t.Add("Value", "text").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("ServerProperty_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("serverstats")
-	t.Add("StatDate", "datetime").NotNullable()
-	t.Add("Clients", "int(11)").NotNullable()
-	t.Add("CPU", "double").NotNullable()
-	t.Add("Upload", "int(11)").NotNullable()
-	t.Add("Download", "int(11)").NotNullable()
-	t.Add("Memory", "bigint(20)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("serverstats_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("SinglePermission")
-	t.AddWithIndex("PlayerID", "varchar(255)").NotNullable()
-	t.AddWithIndex("Command", "varchar(255)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("SinglePermission_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Specialization")
-	t.AddPrimary("SpecializationID", "int(11)").NotNullable()
-	t.AddUnique("KeyName", "varchar(100)").NotNullable()
-	t.Add("Name", "varchar(255)").NotNullable()
-	t.Add("Icon", "smallint(5) unsigned").NotNullable()
-	t.Add("Description", "text")
-	t.Add("Implementation", "varchar(255)")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 225
-	allTables = append(allTables, *t)
-
-	t = NewTable("SpecXAbility")
-	t.AddPrimary("SpecXabilityID", "int(11)").NotNullable()
-	t.AddWithIndex("Spec", "varchar(100)").NotNullable()
-	t.Add("SpecLevel", "int(11)").NotNullable()
-	t.AddWithIndex("AbilityKey", "varchar(100)").NotNullable()
-	t.Add("AbilityLevel", "int(11)").NotNullable()
-	t.Add("ClassId", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 367
-	allTables = append(allTables, *t)
-
-	t = NewTable("SpellLine")
-	t.AddPrimary("SpellLineID", "int(11)").NotNullable()
-	t.AddUnique("KeyName", "varchar(255)").NotNullable()
-	t.Add("Name", "varchar(255)")
-	t.AddWithIndex("Spec", "varchar(100)")
-	t.Add("IsBaseLine", "tinyint(1)").NotNullable()
-	t.Add("ClassIDHint", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1463
-	allTables = append(allTables, *t)
-
-	t = NewTable("Spell")
-	t.AddUnique("SpellID", "int(11)").NotNullable()
-	t.Add("ClientEffect", "int(11)").NotNullable()
-	t.Add("Icon", "int(11)").NotNullable()
-	t.Add("Name", "text").NotNullable()
-	t.Add("Description", "text").NotNullable()
-	t.Add("Target", "text").NotNullable()
-	t.Add("Range", "int(11)").NotNullable()
-	t.Add("Power", "int(11)").NotNullable()
-	t.Add("CastTime", "double").NotNullable()
-	t.Add("Damage", "double").NotNullable()
-	t.Add("DamageType", "int(11)").NotNullable()
-	t.Add("Type", "text")
-	t.Add("Duration", "int(11)").NotNullable()
-	t.Add("Frequency", "int(11)").NotNullable()
-	t.Add("Pulse", "int(11)").NotNullable()
-	t.Add("PulsePower", "int(11)").NotNullable()
-	t.Add("Radius", "int(11)").NotNullable()
-	t.Add("RecastDelay", "int(11)").NotNullable()
-	t.Add("ResurrectHealth", "int(11)").NotNullable()
-	t.Add("ResurrectMana", "int(11)").NotNullable()
-	t.Add("Value", "double").NotNullable()
-	t.Add("Concentration", "int(11)").NotNullable()
-	t.Add("LifeDrainReturn", "int(11)").NotNullable()
-	t.Add("AmnesiaChance", "int(11)").NotNullable()
-	t.Add("Message1", "text")
-	t.Add("Message2", "text")
-	t.Add("Message3", "text")
-	t.Add("Message4", "text")
-	t.Add("InstrumentRequirement", "int(11)").NotNullable()
-	t.Add("SpellGroup", "int(11)").NotNullable()
-	t.Add("EffectGroup", "int(11)").NotNullable()
-	t.Add("SubSpellID", "int(11)").NotNullable()
-	t.Add("MoveCast", "tinyint(1)").NotNullable()
-	t.Add("Uninterruptible", "tinyint(1)").NotNullable()
-	t.Add("IsFocus", "tinyint(1)").NotNullable()
-	t.Add("SharedTimerGroup", "int(11)").NotNullable()
-	t.Add("IsPrimary", "tinyint(1)").NotNullable()
-	t.Add("IsSecondary", "tinyint(1)").NotNullable()
-	t.Add("AllowBolt", "tinyint(1)").NotNullable()
-	t.Add("PackageID", "text")
-	t.Add("TooltipId", "smallint(5) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Spell_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("SpellXCustomValues")
-	t.AddWithIndex("SpellID", "int(11)").NotNullable()
-	t.AddWithIndex("KeyName", "varchar(100)").NotNullable()
-	t.Add("Value", "varchar(255)")
-	t.AddPrimary("CustomParamID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 1
-	allTables = append(allTables, *t)
-
-	t = NewTable("StarterEquipment")
-	t.AddPrimary("StarterEquipmentID", "int(11)").NotNullable()
-	t.Add("Class", "text").NotNullable()
-	t.Add("TemplateID", "varchar(255)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 128
-	allTables = append(allTables, *t)
-
-	t = NewTable("StartupLocation")
-	t.AddPrimary("StartupLoc_ID", "int(11)").NotNullable()
-	t.Add("XPos", "int(11)").NotNullable()
-	t.Add("YPos", "int(11)").NotNullable()
-	t.Add("ZPos", "int(11)").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.Add("Region", "int(11)").NotNullable()
-	t.Add("MinVersion", "int(11)").NotNullable()
-	t.Add("RealmID", "int(11)").NotNullable()
-	t.Add("RaceID", "int(11)").NotNullable()
-	t.Add("ClassID", "int(11)").NotNullable()
-	t.Add("ClientRegionID", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 137
-	allTables = append(allTables, *t)
-
-	t = NewTable("Style")
-	t.AddPrimary("StyleID", "int(11)").NotNullable()
-	t.Add("ID", "int(11)").NotNullable()
-	t.Add("ClassId", "int(11)").NotNullable()
-	t.Add("Name", "varchar(255)").NotNullable()
-	t.AddWithIndex("SpecKeyName", "varchar(100)").NotNullable()
-	t.Add("SpecLevelRequirement", "int(11)").NotNullable()
-	t.Add("Icon", "int(11)").NotNullable()
-	t.Add("EnduranceCost", "int(11)").NotNullable()
-	t.Add("StealthRequirement", "tinyint(1)").NotNullable()
-	t.Add("OpeningRequirementType", "int(11)").NotNullable()
-	t.Add("OpeningRequirementValue", "int(11)").NotNullable()
-	t.Add("AttackResultRequirement", "int(11)").NotNullable()
-	t.Add("WeaponTypeRequirement", "int(11)").NotNullable()
-	t.Add("GrowthOffset", "double").NotNullable()
-	t.Add("GrowthRate", "double").NotNullable()
-	t.Add("BonusToHit", "int(11)").NotNullable()
-	t.Add("BonusToDefense", "int(11)").NotNullable()
-	t.Add("TwoHandAnimation", "int(11)").NotNullable()
-	t.Add("RandomProc", "tinyint(1)").NotNullable()
-	t.Add("ArmorHitLocation", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AutoIncrement = 50014
-	allTables = append(allTables, *t)
-
-	t = NewTable("StyleXSpell")
-	t.Add("SpellID", "int(11)").NotNullable()
-	t.Add("ClassID", "int(11)").NotNullable()
-	t.AddWithIndex("StyleID", "int(11)").NotNullable()
-	t.Add("Chance", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("StyleXSpell_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Task")
-	t.AddPrimary("Character_ID", "varchar(255)").NotNullable()
-	t.Add("TimeOut", "datetime")
-	t.Add("TaskType", "text")
-	t.Add("TasksDone", "int(11)").NotNullable()
-	t.Add("CustomPropertiesString", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Task_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("Teleport")
-	t.Add("Type", "text").NotNullable()
-	t.AddWithIndex("TeleportID", "varchar(255)").NotNullable()
-	t.Add("Realm", "int(11)").NotNullable()
-	t.Add("RegionID", "int(11)").NotNullable()
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Heading", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("Teleport_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("TownCrierMessages")
-	t.AddPrimary("TownCrierID", "varchar(255)").NotNullable()
-	t.Add("Message", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("TownCrierMessages_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	t = NewTable("WorldObject")
-	t.Add("ClassType", "text")
-	t.Add("TranslationId", "text")
-	t.Add("Name", "text").NotNullable()
-	t.Add("ExamineArticle", "text")
-	t.Add("X", "int(11)").NotNullable()
-	t.Add("Y", "int(11)").NotNullable()
-	t.Add("Z", "int(11)").NotNullable()
-	t.Add("Heading", "smallint(5) unsigned").NotNullable()
-	t.AddWithIndex("Region", "smallint(5) unsigned").NotNullable()
-	t.Add("Model", "smallint(5) unsigned").NotNullable()
-	t.Add("Emblem", "int(11)").NotNullable()
-	t.Add("Realm", "tinyint(3) unsigned").NotNullable()
-	t.Add("RespawnInterval", "int(11)").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("WorldObject_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("ZonePoint")
-	t.AddWithIndex("Id", "smallint(5) unsigned").NotNullable()
-	t.Add("TargetX", "int(11)").NotNullable()
-	t.Add("TargetY", "int(11)").NotNullable()
-	t.Add("TargetZ", "int(11)").NotNullable()
-	t.Add("TargetRegion", "smallint(5) unsigned").NotNullable()
-	t.Add("TargetHeading", "smallint(5) unsigned").NotNullable()
-	t.Add("SourceX", "int(11)").NotNullable()
-	t.Add("SourceY", "int(11)").NotNullable()
-	t.Add("SourceZ", "int(11)").NotNullable()
-	t.Add("SourceRegion", "smallint(5) unsigned").NotNullable()
-	t.AddWithIndex("Realm", "smallint(5) unsigned").NotNullable()
-	t.Add("ClassType", "text")
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddPrimary("ZonePoint_ID", "varchar(255)").NotNullable()
-	allTables = append(allTables, *t)
-
-	t = NewTable("Zones")
-	t.AddPrimary("ZoneID", "int(11)").NotNullable()
-	t.AddWithIndex("RegionID", "smallint(5) unsigned").NotNullable()
-	t.Add("Name", "text").NotNullable()
-	t.Add("IsLava", "tinyint(1)").NotNullable()
-	t.Add("DivingFlag", "tinyint(3) unsigned").NotNullable()
-	t.Add("WaterLevel", "int(11)").NotNullable()
-	t.Add("OffsetY", "int(11)").NotNullable()
-	t.Add("OffsetX", "int(11)").NotNullable()
-	t.Add("Width", "int(11)").NotNullable()
-	t.Add("Height", "int(11)").NotNullable()
-	t.Add("Experience", "int(11)").NotNullable()
-	t.Add("Realmpoints", "int(11)").NotNullable()
-	t.Add("Bountypoints", "int(11)").NotNullable()
-	t.Add("Coin", "int(11)").NotNullable()
-	t.Add("Realm", "tinyint(3) unsigned").NotNullable()
-	t.Add("LastTimeRowUpdated", "datetime").NotNullable()
-	t.AddUnique("Zones_ID", "varchar(255)")
-	allTables = append(allTables, *t)
-
-	return allTables
-}
-
-var (
-	cachedStaticTables []string
-)
-
-func GetStaticTableNames() []string {
-	if cachedStaticTables != nil {
-		return cachedStaticTables
-	}
-	jsonPath := filepath.Join(utils.RootFolder(), "data", "_StaticTables.json")
-	content, err := os.ReadFile(jsonPath)
-	if err != nil {
-		panic(err)
-	}
-	json.Unmarshal(content, &cachedStaticTables)
-	return cachedStaticTables
-}
-
-func IsTableStatic(tableName string) bool {
-	return slices.Contains(GetStaticTableNames(), tableName)
-}
